testability refactor
