+++ test/ghrequestorTests.js
@@ -2,32 +2,33 @@ const assert = require('chai').assert;
 const chai = require('chai');
 const expect = require('chai').expect;
 const extend = require('extend');
-const requestor = require('../lib/ghrequestor.js');
+const requestor = require('../lib/requestor.js');
+const GHRequestor = require('../lib/ghrequestor.js')
 const request = require('requestretry');
 
 const urlHost = 'https://test.com';
 
-// describe('Request option merging', () => {
-//   it('should merge and override properties', () => {
-//     const result = new requestor.GHRequestor({
-//       retryDelay: 10,
-//       testProperty: 'test value'
-//     });
-//     expect(result.options.retryDelay).to.equal(10);
-//     expect(result.options.testProperty).to.equal('test value');
-//   });
+describe('Request option merging', () => {
+  it('should merge and override properties', () => {
+    const result = new GHRequestor({
+      retryDelay: 10,
+      testProperty: 'test value'
+    });
+    expect(result.options.retryDelay).to.equal(10);
+    expect(result.options.testProperty).to.equal('test value');
+  });
 
-//   it('should merge and override headers', () => {
-//     const result = new requestor.GHRequestor({
-//       headers: {
-//         'User-Agent': 'test agent',
-//         authorization: 'test auth'
-//       }
-//     });
-//     expect(result.options.headers['User-Agent']).to.equal('test agent');
-//     expect(result.options.headers.authorization).to.equal('test auth');
-//   });
-// });
+  it('should merge and override headers', () => {
+    const result = new GHRequestor({
+      headers: {
+        'User-Agent': 'test agent',
+        authorization: 'test auth'
+      }
+    });
+    expect(result.options.headers['User-Agent']).to.equal('test agent');
+    expect(result.options.headers.authorization).to.equal('test auth');
+  });
+});
 
 describe('Request retry and success', () => {
   it('should be able to get a single page resource', () => {
